wip audio

+++ extern/musyx/include/musyx/sal.h
@@ -69,7 +69,7 @@ void salCalcVolume(u8 voltab_index, SAL_VOLINFO* vi, f32 vol, u32 pan, u32 span,
 void salReconnectVoice(DSPvoice* dsp_vptr, u8 studio);
 void* salMalloc(u32 len);
 void salFree(void* addr);
-void salBuildCommandList(signed short* dest, unsigned long nsDelay);
+void salBuildCommandList(s16* dest, u32 nsDelay);
 void salStartDsp(u16* cmdList);
 void salCtrlDsp(s16* dest);
 u32 salSynthSendMessage(DSPvoice* dsp_vptr, u32 mesg);
